[1.1.0.3]Tableの書き方を更新したけど、まだ実験していない

+++ Assets/Resources/DenQ_SweeperScript/Table/DataStruct/BombData.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 public class BombData : ItemBase{
 	public string	name;
 	public uint 	bombType;
-	public int		level;
-	public int		hp;
+	public uint		level;
+	public uint		hp;
 	public float	time;
-	public float	damage;
+	public uint		damage;
 	public float	damageRange;
 
 }
